requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --no-deps -o requirements.txt pyproject.toml
azure-search-documents==11.5.2
    # via flock-core (pyproject.toml)
chromadb==0.6.3
    # via flock-core (pyproject.toml)
cloudpickle==3.1.1
    # via flock-core (pyproject.toml)
datasets==3.5.0
    # via flock-core (pyproject.toml)
devtools==0.12.2
    # via flock-core (pyproject.toml)
dspy==2.6.16
    # via flock-core (pyproject.toml)
duckduckgo-search==7.3.2
    # via flock-core (pyproject.toml)
fastapi==0.115.12
    # via flock-core (pyproject.toml)
httpx==0.28.1
    # via flock-core (pyproject.toml)
litellm==1.63.7
    # via flock-core (pyproject.toml)
loguru==0.7.3
    # via flock-core (pyproject.toml)
matplotlib==3.10.1
    # via flock-core (pyproject.toml)
msgpack==1.1.0
    # via flock-core (pyproject.toml)
nltk==3.9.1
    # via flock-core (pyproject.toml)
notion-client==2.3.0
    # via flock-core (pyproject.toml)
opentelemetry-api==1.30.0
    # via flock-core (pyproject.toml)
opentelemetry-exporter-jaeger==1.21.0
    # via flock-core (pyproject.toml)
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
    # via flock-core (pyproject.toml)
opentelemetry-exporter-otlp==1.30.0
    # via flock-core (pyproject.toml)
opentelemetry-instrumentation-logging==0.51b0
    # via flock-core (pyproject.toml)
opentelemetry-sdk==1.30.0
    # via flock-core (pyproject.toml)
pandas==2.2.3
    # via flock-core (pyproject.toml)
pillow==11.1.0
    # via flock-core (pyproject.toml)
prometheus-client==0.21.1
    # via flock-core (pyproject.toml)
psutil==7.0.0
    # via flock-core (pyproject.toml)
pydantic==2.10.6
    # via flock-core (pyproject.toml)
pydantic-settings==2.8.1
    # via flock-core (pyproject.toml)
python-box==7.3.2
    # via flock-core (pyproject.toml)
python-decouple==3.8
    # via flock-core (pyproject.toml)
python-dotenv==1.1.0
    # via flock-core (pyproject.toml)
python-fasthtml==0.12.12
    # via flock-core (pyproject.toml)
pyyaml==6.0.2
    # via flock-core (pyproject.toml)
questionary==2.1.0
    # via flock-core (pyproject.toml)
rich==13.9.4
    # via flock-core (pyproject.toml)
sentence-transformers==4.0.1
    # via flock-core (pyproject.toml)
temporalio==1.9.0
    # via flock-core (pyproject.toml)
tiktoken==0.9.0
    # via flock-core (pyproject.toml)
toml==0.10.2
    # via flock-core (pyproject.toml)
tqdm==4.67.1
    # via flock-core (pyproject.toml)
uvicorn==0.34.0
    # via flock-core (pyproject.toml)
zep-python==2.0.2
    # via flock-core (pyproject.toml)
